unify features, rename feature_creation to feature_extraction, tda_features to topology
