--- giottotime/feature_creation/standard_features/__init__.py
@@ -1,9 +1,0 @@
-from .base import StandardFeature
-from .standard_features import PeriodicSeasonalFeature, ConstantFeature, CustomFeature
-
-__all__ = [
-    "StandardFeature",
-    "PeriodicSeasonalFeature",
-    "ConstantFeature",
-    "CustomFeature",
-]
