+++ giottotime/feature_extraction/__init__.py
@@ -1,4 +1,9 @@
-from .base import IndexDependentFeature
+"""
+The :mod:`giottotime.feature_extraction` module deals with the creation of features
+starting from a time series.
+"""
+
+from .feature_creation import FeatureCreation
 from .calendar_features import CalendarFeature
 from .time_series_features import (
     ShiftFeature,
@@ -6,25 +11,28 @@ from .time_series_features import (
     PolynomialFeature,
     ExogenousFeature,
 )
+
+from .topology import AmplitudeFeature, AvgLifeTimeFeature, \
+    BettiCurvesFeature, NumberOfRelevantHolesFeature
+
+from .standard_features import ConstantFeature, PeriodicSeasonalFeature, \
+    CustomFeature
+
 from .trend_features import (
     DetrendedFeature,
     RemovePolynomialTrend,
-    RemoveExponentialTrend,
-)
-from .tda_features import (
-    AmplitudeFeature,
-    AvgLifeTimeFeature,
-    BettiCurvesFeature,
-    NumberOfRelevantHolesFeature,
-)
+    RemoveExponentialTrend)
 
 __all__ = [
-    "IndexDependentFeature",
-    "CalendarFeature",
+    "FeatureCreation",
     "ShiftFeature",
     "MovingAverageFeature",
+    "ConstantFeature",
     "PolynomialFeature",
     "ExogenousFeature",
+    "topology",
+    "CalendarFeature",
+    "PeriodicSeasonalFeature",
     "DetrendedFeature",
     "RemovePolynomialTrend",
     "RemoveExponentialTrend",
