+++ giottotime/feature_extraction/tests/test_feature_creation.py
@@ -5,8 +5,8 @@ import numpy as np
 from hypothesis._settings import duration
 
 from giottotime.utils.hypothesis.time_indexes import giotto_time_series
-from giottotime.feature_creation import ShiftFeature, MovingAverageFeature
-from giottotime.feature_creation.feature_creation import (
+from giottotime.feature_extraction import ShiftFeature, MovingAverageFeature
+from giottotime.feature_extraction.feature_creation import (
     _check_feature_names,
     FeatureCreation,
 )
