+++ giottotime/feature_extraction/tests/test_standard_features.py
@@ -5,7 +5,7 @@ from hypothesis import given, strategies as st
 from pandas.util import testing as testing
 
 from giottotime.utils.hypothesis.time_indexes import giotto_time_series
-from giottotime.feature_creation import (
+from giottotime.feature_extraction import (
     ConstantFeature,
     CustomFeature,
     PeriodicSeasonalFeature,
