+++ giottotime/feature_extraction/tests/test_time_series_features.py
@@ -6,7 +6,7 @@ from hypothesis import given, strategies as st
 from sklearn.preprocessing import PolynomialFeatures
 
 from giottotime.utils.hypothesis.time_indexes import giotto_time_series
-from giottotime.feature_creation import (
+from giottotime.feature_extraction import (
     ShiftFeature,
     MovingAverageFeature,
     ExogenousFeature,
