+++ giottotime/feature_extraction/topology/tests/test_amplitude_feature.py
@@ -1,9 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from giottotime.feature_creation.index_dependent_features.tda_features import (
-    AmplitudeFeature,
-)
+from giottotime.feature_extraction.topology import AmplitudeFeature
 
 
 def test_correct_amplitude_computation():
