+++ giottotime/feature_extraction/topology/tests/test_base.py
@@ -6,9 +6,7 @@ import pandas as pd
 import pytest
 from hypothesis import given, strategies as st
 
-from giottotime.feature_creation.index_dependent_features.tda_features import (
-    TDAFeatures,
-)
+from giottotime.feature_extraction.topology import TDAFeatures
 
 
 class BaseTDAFeature(TDAFeatures):
