+++ giottotime/feature_extraction/topology/tests/test_betti_curves_feature.py
@@ -2,9 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from giottotime.feature_creation.index_dependent_features.tda_features import (
-    BettiCurvesFeature,
-)
+from giottotime.feature_extraction.topology import BettiCurvesFeature
 
 
 def test_correct_betti_curves_mean():
