+++ giottotime/feature_extraction/topology/tests/test_relevant_holes_feature.py
@@ -3,9 +3,7 @@ import pandas as pd
 import pytest
 from hypothesis import given, strategies as st
 
-from giottotime.feature_creation.index_dependent_features.tda_features import (
-    NumberOfRelevantHolesFeature,
-)
+from giottotime.feature_extraction.topology import NumberOfRelevantHolesFeature
 
 
 @given(st.integers(min_value=3))
