+++ giottotime/models/__init__.py
@@ -1,5 +1,5 @@
 """
-The :mod:`giottotime.feature_creation` module contains a collection of different kind
+The :mod:`giottotime.feature_extraction` module contains a collection of different kind
 of machine learning models, for dealing with time series and non time series data.
 """
